# docs/openapi.json
{
 "components": {
  "schemas": {
   "Body_install_plugin_api_v1_plugins_install_post": {
    "properties": {
     "file": {
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "type": "string"
     },
     "signature": {
      "anyOf": [
       {
        "contentMediaType": "application/octet-stream",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Signature"
     }
    },
    "required": [
     "file"
    ],
    "title": "Body_install_plugin_api_v1_plugins_install_post",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PluginBuildListResponse": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PluginBuildResponse"
      },
      "title": "Items",
      "type": "array"
     },
     "total": {
      "default": 0,
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items"
    ],
    "title": "PluginBuildListResponse",
    "type": "object"
   },
   "PluginBuildResponse": {
    "properties": {
     "build_id": {
      "title": "Build Id",
      "type": "string"
     },
     "error_summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Summary"
     },
     "package_sha256": {
      "title": "Package Sha256",
      "type": "string"
     },
     "plugin_id": {
      "title": "Plugin Id",
      "type": "string"
     },
     "runtime_fingerprint": {
      "title": "Runtime Fingerprint",
      "type": "string"
     },
     "runtime_type": {
      "enum": [
       "conda-pack",
       "docker"
      ],
      "title": "Runtime Type",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "target_arch": {
      "title": "Target Arch",
      "type": "string"
     },
     "target_os": {
      "title": "Target Os",
      "type": "string"
     },
     "version": {
      "title": "Version",
      "type": "string"
     }
    },
    "required": [
     "build_id",
     "plugin_id",
     "version",
     "runtime_type",
     "target_os",
     "target_arch",
     "status",
     "package_sha256",
     "runtime_fingerprint"
    ],
    "title": "PluginBuildResponse",
    "type": "object"
   },
   "PluginDetailResponse": {
    "properties": {
     "author": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Author"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "latest_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latest Version"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "versions": {
      "items": {
       "$ref": "#/components/schemas/PluginVersionDetail"
      },
      "title": "Versions",
      "type": "array"
     }
    },
    "required": [
     "id",
     "name",
     "versions"
    ],
    "title": "PluginDetailResponse",
    "type": "object"
   },
...
   "PluginVersionDetail": {
    "properties": {
     "manifest": {
      "additionalProperties": true,
      "title": "Manifest",
      "type": "object"
     },
     "sdk_version": {
      "title": "Sdk Version",
      "type": "string"
     },
     "source_sha256": {
      "title": "Source Sha256",
      "type": "string"
     },
     "spec_version": {
      "title": "Spec Version",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "version": {
      "title": "Version",
      "type": "string"
     }
    },
    "required": [
     "version",
     "spec_version",
     "sdk_version",
     "source_sha256",
     "status",
     "manifest"
    ],
    "title": "PluginVersionDetail",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "Python Service Hub",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/plugin-builds": {
   "get": {
    "description": "Return a page of plugin Builds newest-first plus the total matching the filters.\n\nThe default limit stays at 100 (not the 50 used by /jobs) on purpose: this\nendpoint predates pagination and returned up to 100 Builds, so parameterless\ncallers keep seeing the same newest-100 window.",
    "operationId": "list_plugin_builds_api_v1_plugin_builds_get",
    "parameters": [
     {
      "in": "query",
      "name": "plugin_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 64,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Plugin Id"
      }
     },
     {
      "in": "query",
      "name": "runtime_type",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "conda-pack",
          "docker"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Runtime Type"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 200,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildListResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Plugin Builds",
    "tags": [
     "plugins"
    ]
   }
  },
  "/api/v1/plugin-builds/{build_key}": {
   "delete": {
    "description": "Delete one DEPRECATED Build after the full guard chain (review doc B4b).\n\nAdmin-only and irreversible. Guards (any hit -> 409): the Build must be\nDEPRECATED (deprecate is the mandatory cooling-off step); no enabled\nSchedule may still resolve the (plugin, version, runtime) triple as its\nlast Build; no Pipeline step may reference it; no active Job (PENDING /\nPREPARING / RUNNING / CANCEL_REQUESTED) may reference it. Deletion order\nfollows R3: registry rows and storage first, the docker image second —\nnever any form of prune.",
    "operationId": "delete_plugin_build_api_v1_plugin_builds__build_key__delete",
    "parameters": [
     {
      "in": "path",
      "name": "build_key",
      "required": true,
      "schema": {
       "title": "Build Key",
       "type": "string"
      }
     },
     {
      "description": "演练模式: 只返回将要执行的动作清单, 不做任何变更",
      "in": "query",
      "name": "dry_run",
      "required": false,
      "schema": {
       "default": false,
       "description": "演练模式: 只返回将要执行的动作清单, 不做任何变更",
       "title": "Dry Run",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Delete Plugin Build Api V1 Plugin Builds  Build Key  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Plugin Build",
    "tags": [
     "plugins"
    ]
   },
   "get": {
    "description": "Return one Plugin Build summary.",
    "operationId": "get_build_api_v1_plugin_builds__build_key__get",
    "parameters": [
     {
      "in": "path",
      "name": "build_key",
      "required": true,
      "schema": {
       "title": "Build Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Build",
    "tags": [
     "plugins"
    ]
   }
  },
  "/api/v1/plugin-builds/{build_key}/deprecate": {
   "post": {
    "description": "Deprecate one READY or ENABLED Build; the transition is one-way.\n\nA DEPRECATED Build can no longer back new Jobs (job resolution only accepts\nENABLED Builds) and is marked as pending future cleanup, while staying\nvisible in the registry with its storage and environment material intact.\nThere is deliberately no revival path: ``enable`` refuses a DEPRECATED\nBuild, and re-installing the exact same version is rejected by the\nimmutable Build uniqueness constraint, so the only way to serve this\nplugin version again is to install its package under a new version and\nenable the fresh Build (see review doc G4(1), B4a).",
    "operationId": "deprecate_build_api_v1_plugin_builds__build_key__deprecate_post",
    "parameters": [
     {
      "in": "path",
      "name": "build_key",
      "required": true,
      "schema": {
       "title": "Build Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Deprecate Build",
    "tags": [
     "plugins"
    ]
   }
  },
  "/api/v1/plugin-builds/{build_key}/disable": {
   "post": {
    "description": "Disable a Build when it has no active Jobs.",
    "operationId": "disable_build_api_v1_plugin_builds__build_key__disable_post",
    "parameters": [
     {
      "in": "path",
      "name": "build_key",
      "required": true,
      "schema": {
       "title": "Build Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Disable Build",
    "tags": [
     "plugins"
    ]
   }
  },
  "/api/v1/plugin-builds/{build_key}/enable": {
   "post": {
    "description": "Enable one READY Build explicitly.",
    "operationId": "enable_build_api_v1_plugin_builds__build_key__enable_post",
    "parameters": [
     {
      "in": "path",
      "name": "build_key",
      "required": true,
      "schema": {
       "title": "Build Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Enable Build",
    "tags": [
     "plugins"
    ]
   }
  },
...
  "/api/v1/plugins/install": {
   "post": {
    "description": "Verify and register one uploaded `.pypkg` Build package.\n\nThe optional ``signature`` multipart field carries the ``.sig`` document\nproduced by ``hub-plugin sign`` (B9). Whether it is required, verified, or\nignored is decided by ``plugins.signature``: with no public keys configured\nthe endpoint behaves exactly as before B9.",
    "operationId": "install_plugin_api_v1_plugins_install_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_install_plugin_api_v1_plugins_install_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginBuildResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Install Plugin",
    "tags": [
     "plugins"
    ]
   }
  },
  "/api/v1/plugins/{plugin_id}": {
   "get": {
    "description": "Return one installed plugin and the public manifests of its versions.",
    "operationId": "get_plugin_detail_api_v1_plugins__plugin_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "plugin_id",
      "required": true,
      "schema": {
       "title": "Plugin Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PluginDetailResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Plugin Detail",
    "tags": [
     "plugins"
    ]
   }
  },
...
 }
}